% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{blend}
\alias{blend}
\title{Image blending}
\usage{
blend(a, b, mode = c("screen", "multiply"))
}
\arguments{
\item{a}{png-array, as produced by \code{\link[png]{readPNG}}.}

\item{b}{png-array.}

\item{mode}{Blend mode.}
}
\value{
A blended png-array.
}
\description{
Image blending
}
\seealso{
mono2png
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
