% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{mono2hex}
\alias{mono2hex}
\title{Convert monochrome matrix to hex color matrix}
\usage{
mono2hex(x, pal = c("#559988", "#ff9933"))
}
\arguments{
\item{x}{Monochrome matrix (numeric with all elements in [0,1]).}

\item{pal}{Color palette.}
}
\value{
A character matrix of hexadecimal color codes.
}
\description{
Convert monochrome matrix to hex color matrix
}
\seealso{
mono2png
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
