% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot3d.adeba}
\alias{plot3d.adeba}
\title{Plot adeba estimate in 3d}
\usage{
\method{plot3d}{adeba}(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Sent to \code{\link[rgl]{plot3d}} of the \code{rgl} package.}
}
\description{
Using OpenGL. Re-exported from the \code{rgl} package, so please see
\code{\link[rgl]{plot3d}} for details.
}
\examples{
x.train <- 3 + sweep(matrix(rnorm(60), 30), 2, 1:2, "*") \%*\% matrix(c(1, .4, .4, 1), 2)
f <- adeba(x.train, adaptive=FALSE)

library(rgl)
plot3d(f)
}
